# Less clock_gettime calls