I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of fast queries when `max_execution_time=0` by reducing the number of `clock_gettime` system calls.

Detailed description / Documentation draft:

```
-- before 
SELECT count() FROM zeros(100000000000)

┌──────count()─┐
│ 100000000000 │
└──────────────┘

1 rows in set. Elapsed: 11.320 sec. Processed 100.00 billion rows, 100.00 GB (8.83 billion rows/s., 8.83 GB/s.)


-- after 
select count() from zeros(100000000000);

┌──────count()─┐
│ 100000000000 │
└──────────────┘

1 rows in set. Elapsed: 4.627 sec. Processed 100.00 billion rows, 100.00 GB (21.61 billion rows/s., 21.61 GB/s.)
```