@filimonov 2021-08-07
Performance tests don't show any difference because of https://github.com/ClickHouse/ClickHouse/blob/e1927118cd1d020099bbd564a8b23cf4c5bc5c40/docker/test/performance-comparison/config/users.d/perf-comparison-tweaks-users.xml#L16