@filimonov 2021-08-07
Yandex synchronization - maybe i should overload the function instead of changing its interface?