@filimonov 2021-08-09
BTW: it looks like part of the original issue is build-related, on official builds difference is smaller, while still clearly visible:
```
SET max_execution_time = 0; SELECT count() FROM zeros_mt(1000000000000);

1 rows in set. Elapsed: 10.922 sec. Processed 1.00 trillion rows, 1.00 TB (91.56 billion rows/s., 91.56 GB/s.)

SET max_execution_time = 300; SELECT count() FROM zeros_mt(1000000000000);

1 rows in set. Elapsed: 13.124 sec. Processed 1.00 trillion rows, 1.00 TB (76.20 billion rows/s., 76.20 GB/s.)
```