@alexey-milovidov 2021-08-09
Why do you even try to use unofficial builds?
Only official builds are supported. And we give no guarantee about any third-party builds.
Moreover we are confident that third-party builds will work terribly in multiple ways, because they are not covered by our continuous integration, including performance and fuzz testing.